Move `Route` to `Map` folder
