+++ appOPHD/UI/MiniMap.cpp
@@ -1,10 +1,10 @@
 #include "MiniMap.h"
 
 #include "../Cache.h"
+#include "../Map/Route.h"
 #include "../Map/TileMap.h"
 #include "../Map/MapView.h"
 #include "../MapObjects/Robot.h"
-#include "../States/Route.h"
 #include "../StructureManager.h"
 
 #include <libOPHD/MapObjects/OreDeposit.h>
